refactor: UserApiController 및 테스트 코드 경로 변경
